Base code for WBGT algorithms
Based code for the Liljegren, Dimiceli, and Bernard
wet bulb globe temperature alogithms are being worked on.
Main focus was to get the Liljegren C-code wrapped in Cython/Python
with some APIs for useful functions.

More testing is required of the Liljegren code with others to
follow as may need to use some of the functions from Liljegren
in other algorthims due to incomplete documention.

v0.0.0