Reorganizes and standardize bernard method
Various functions in the bernard method file have been
rearranged so that they appear in a more logical order when
looking through the code.

Docstrings and comments for the various functions have been
greatly improved.

The return of the main function has been updated to a
standarized dict that will be consistent across the various
methods.