Performance improvement to Liljegren C code?
Made some minor formatting changes to the Twb() function in the
underlying C code for the liljegren method that entailed lining
up equals signs to improve readability.

I also slightly modified the code in a way that I think will make
it a little more performant. In the previous (original) version,
the influence of solar radiation on the wet bulb temperature was
calculated within the do loop no matter if it were being considered
or not. Then, this influence was include/excluded by multiplying
it by the value of rad, which was one (1) when including and
zero (0) when excluding solar influence. This ment a lot of
calculations were occuring every loop that may not have been
needed. So, I changed the code so that the 'heat' influence is
zero before looping starts. Inside the loop, a check is now made
to see if rad == 1 (i.e., solar influence is to be included).
If this is true, then the influence is calculated and the 'heat'
variable is updated. Note that 'heat' is always added to the
resultant temperature, but it will only be non-zero IF rad == 1 as
'heat = 0.0' before the looping starts and is not changed unless
rad == 1.

I have also added source code for the  National Renewable Energy
Laboratory (NREL) Solar Position Algorithm (SPA) to the src
directory as I may use it in the future. The header file is also
now imported in the liljegren.c file.

The SPA code may be used in the future because the solar position
algorithm include in the liljegren code is only valid until 2050,
when the some what crude algorithm will begin to have large, and
growing, errors.