API changes, bug fixe, and standardize liljegren
There was a bug in the globeTemperature function where the input
to the function is expected to be in degree Celsius, but the
underlying C function expects temperature in Kelvin. So, 273.15 is
added to the input value before it is passed to the C code.

Have made some major changes to the liljegren function, the biggest
being changing input variables/locations and changing the output
to be the new standardized dict for all algorithms.
For inputs, they have be standardized to match those of the other
algorithms, with 'special' inputs such as the GMT offset, averaging
time for data values, etc. being changed to keyword arguments.
If these are not input by the user, then default values are used.

Also fixed a bug where the estimated speed output by the function was
not updated correctly.