Adds a bunch of code for algorithm validation
The added code is used to generate plots from ECONet data to
'validate' the various WBGT algorithms. The plotting
codes are all very similar, with the difference being what variable
is used to color the symbols. These symbol colors are used to
indicate the influence of a given variable on the algorithm.

For example, all the plots show the calculated WBGT versus the
measured WBGT by ECONet. The points are then colored by solar
irradiance values to give an idea of how solar irrandiance is
influencing differences between the algorithms and truth