Unit tests, NREL SPA, reorganize
Starts to add unit test code for modules

Adds support for the NREL Solar Position Algorithm (SPA)
to the Liljegren code. This is useful for solar position
information as the Liljegren solar position code is only valid
from 1950-2050, where has NREL SPA is valid from -2000-6000.
More testing of this implementation is required.

Reorganizes some testing code to be in the 'validation' module.

Adds some docstrings to modules, ensures docstrings are consistent,
ensures that dicts returned from WBGT functions are consistent.