Adds input for seconds; gets unit tests working
In previous versions it was not possible to input the seconds
part of an observation time into the WBGT functions, namely because
the Liljegren functions did NOT support this option. Upon looking
at the Liljegren calc_solar_position code closely, I believe I
found a bug in their date/time calculations when adjusting for
the observations averaging period.

When adjusting the day based on averaging (and timezone), no checks
were made to ensure the year did not change. For example, if an
observation occured at 00:00:00 on the first of the year, any
adjustmet for averaging places the time in the previous year. I
did not see any compensation for this. Note that I do not think
changing months is an issue as dates are converted to julian day
of the year, so this should not be an issue.

Anyway, the inclusion of the NREL SPA code created an issue where
seconds could be input into that funcion, so it was important to
make that change.

Now, date/time manipulation is done before calling either the
Liljegren or SPA code (i.e., timezone offsets and averaging windows)
so that everything is consistent.

Have updated all the validation code to support seconds inputs,
although has not been tested.

The unit test code has been tested and is working