Gets Bernard method working, linting
Performed a lot of linting of code to fix formatting.
Adds lots of docstrings. Need to do one more look
through to ensure function calls are consistent.

Gets the Bernard method working with some help from
the Liljegren code. All the formulas from the Bernard
paper are used, with some influence from Liljegren for
the heat input into for the globe temperature. Note
that these are NOT exactly the same equations, with the
Bernard being slightly simplified.

Consolidates the psychrometric wetbulb code into a single
cython program. This includes the Stull and Iribarne
methods.