Improves plotting, adds some I/O
Plotting has been improved through the addition of a thinning
function to reduce number of points (and size of EPS files).
Have also improved the plot_by_hue function to be a little cleaner.

Adds the data_io module for writing/reading parquet files with
metadata for the columns of data. The metadata are written into
the scheme for the file and parsed out of the scheme on read.

Creates the stats module for computing some basic statistics on
the data.t