Fix for DatetimeIndex bug, README updates
There was a bug with getting datetimes to work easily for enduser;
may not explicitly be in DatetimeIndex format. Have added a helper
function to convert to DatetimeIndex using the to_datetime()
function from pandas.

Updates README to include information on how to use package with
xarray Datasets/DataArrays.

Updates dimiceli test to use old 'malchaire' algorithm for
natural wetbulb so that test runs