Moves to pyproject installer
Moves from old setup.py based install method to new pyproject.toml
(PEP-621). This took a bit of testing to get the Cython to compile
correctly (still depends on the setup.py file), but have tested on
a macOS (Sonoma) and AWS Linux machine and both work fine.

Following guidance from the Cython documentation, both the source
pyx and the cythonized C files are now included in the distribution.
Providing both files enables users to modify the Cython, re-cythonize
it, and then install updates if they so choose.

However, installation of the package now only depends on the cythonized
files so that the end use does not need to have Cython installed AND so
that files are consistent (i.e., different version of Cython
converted in a different way).
Small changes to some unit test files have been made after finding
that values were every-so-slightly different between macOS and
the AWS Linux instance (15th decimal). Tests, should all pass now.