Updates Dimiceli with min_speed
Discussion of wind speed lead to wanting to force minimum wind
speed to 2 knots as anything less than this value is reported as
calm by ASOS stations. As the Dimiceli method also has a mimimum
wind speed for the calculations, the large of the hard Dimiceli
limit and that entered by the user (which defaults to 2 knots) is
used to clip the winds. This ensures that the value is never below
the limit for the Dimiceli method IF the value is set by the user,
while the default minimum speed is 2 knots.

Have also updated the returned dictionary to include the min_speed
used in the calculations