Updates Liljegren for 2 knot wind min
Took some work, but the cython and source c code for the Liljegren
method have been updated to accept minimum wind speed values for
the computations. This update act similarly to the updat to the
Dimiceli method wherein the user can input a min_speed keyword to
specifiy the minimum speed winds should be clipped to. By default,
this value is 2 knots. There is a secondary check (both in the C
and Cython codes) to ensure that this user input speed minimum is
NOT less than the minimum defined in the liljegren C code (0.13 m/s).

Some redundancies in the liljegren c code for check that wind speed
is above the minimum speed for the algorithm have been removed as
these checks were done in functions called after the initial
adjustment/check for wind speed is complete, making them redundant.