Implements 2 knot wind speed min Bernard
The 2 knot wind speed minimum has been implement in the Bernard
algorithm. One difference is that there is no absolute wind speed
minimum for this algorithm. So, if a user overrides the default
min_speed keyword with, say zero (0), then then wind speeds will
be clipped to zero. By default the min_speed is set to the
MIN_SPEED constant, which is 2 knots