Dimiceli docstring and other updates
Small updates to the Dimiceli algorithms.

The adjust_speed_2m function(s) have been updated so that the
zspeed keyword is now an argument. This value is set in the
main wetbulb_globe function to be 10 meters.

Added datetime_check to the NWS version of the algorithm as it
was missing.

Return values in dictionary have been slightly updated so that
units match those of output from other algorithms. All values are
still unity aware, the 'default' units have just been made
consistent across the package.