Moves to metpy relative humidity
Removes the saturation vapor pressure and relative humidity code
from the calc.py module in favor of the metpy.calc implementations.
This is done to ensure more consistancy with unit checks.
This change led to adjustments in all the algorithm codes; however,
the old implementation of the RH code produced very similar
(nearly identical) values, so no changes to output values has
occurred. Unittests showed no significant changes.

Removes the EPSILON value from the constants.py module has was only
being used by the Bernard method. This value/attribute has been
moved to the bernard.pyx module.

Have added code to the bernard.pyx module to compute the atmospheric
emissivity. This value is used to adjust the influence of dry bulb
temperature in the estimation of black globe temp. Currently, a
surface albedo of zero is being used, and this was raised as an
'issue' by a co author. However, changing surface albedo to a non-zero
value introduced large error. I thought that implementing atmos
emissivity would offset the surface albedo, which it does, but not
enough to balance things out.

Updates dimiceli tests for new RH code