Modifies Dimiceli NWS
From discussion with Blaine Thomas of RCC, the convective heat
transfer coefficent is set to zero for overnight hours (zenith > 87),
and the Tg estimation reduces to (eps * Ta**4)**0.25, where eps is
the emissivitiy of the atmospher and Ta is ambient (dry bulb)
temperature. As the Tg estimation algorithm for Tg assumes a non-zero
transfer coefficient is its derivation/estimation.

Bumps patch version number