Adds unit test for multidim/xarray
Adds new unit test that builds an Xarray Dataset of dummy data
to run through the WBGT algorithms to check for consistent
results.

The Dataset is multidimensional. The dimensions are stacked and
units quantified for the reference run, with each variable
pulled from the stacked/quantified Dataset as a numpy/Quantity.
This is the reference data, with other unit tests ensuring that
computations are correctly.

The full, unmodified Dataset is then fed into the various
algorithms. Results from the Dataset input and the aforementioned
numpy/Quantity input are compared for equality.