Unit tests and Attribute bug fix (#5)
* Unittest for N-Dimensional/Xarray Data (#4)

* Adds unit test for multidim/xarray

Adds new unit test that builds an Xarray Dataset of dummy data
to run through the WBGT algorithms to check for consistent
results.

The Dataset is multidimensional. The dimensions are stacked and
units quantified for the reference run, with each variable
pulled from the stacked/quantified Dataset as a numpy/Quantity.
This is the reference data, with other unit tests ensuring that
computations are correctly.

The full, unmodified Dataset is then fed into the various
algorithms. Results from the Dataset input and the aforementioned
numpy/Quantity input are compared for equality.

* Liniting of the unit tests

---------

Co-authored-by: Kyle R. Wodzicki <kwodzicki@cicsnc.org>

* Fixes bug in min_speed units attr for Xarray

When adding the min_speed information to coordinates in Xarray
Dataset output, I grabbed the val.units attribute, but this
returns a Unit object; need string. So, wrapped call in str()
to fix issue.

---------

Co-authored-by: Kyle R. Wodzicki <kwodzicki@cicsnc.org>