More updates for solar.py module and tests
After installing in a clean environment, unittests were showing
differences of a few hundredths for some of the values.
To test this, I created a new test_solar.py module to check that
cza values from pvlib.spa match those of the new numpy-native
computations. This proved that values are within 12 or so decimal
places of precision of eachother.

Beyond that, I do not see anything that is drastically different
between the codes interms of solar adjustments and fraction of
direct radiation, so these very small differences in cza must
be propagating. I assume that the computations of cza are slighly
more accurate in our version as numpy.sum is used which may be
more accurate than the accumulator scalar that is used in pvlib
and in the old implementation.

Have updated some of the date handling to be more robust to errors,
though need to check this in other pipelines.

Have also forced copy of expanded coordinates when working with
Dataset objects to ensure that the values are contiguous so that
the call to .ravel() does not do weird things. Perhaps shoulds
add in checks/asserstions that must be contiguous.