Updates to handle keywords, datetime fix
Realized that if user inputs data values as keywords, the code for
checking if arguments are DataArrays was completely skipped.
So, added in an ARG_ORDER variable that is used to pop off keywords
in the proper order and append them to the args list.

Also added a bug fix for the min_speed value being converted to
a numpy array without units. In the except when trying to
reshape returned WGBT values, we check if object has 'units' attr,
and if does, then we re-add them after running np.asarray().

We now take the ".values" part of DatetimeIndex objects when
adjusting times.