Implements lazy execution of xarray/dask using map_blocks (#6)
* Updates solar for numpy support

The solar.py module has been updated to natively handle multiple
locations through numpy reshaping and broadcasting mechanics.

Support for numba is removed.

Modifies API call for main wbgt function, moving the method
argument to a keyword so that can support xarray.map_blocks()

* Updates unittests for solar.py changes

Have made some updates to the unittest files as outputs from
solar.py have changed very, very slightly (past 8 decimals).
This lead to small changes in unittest values.

However, upon testing in clean environment, unittests are failing
again, so will be debugging.

* More updates for solar.py module and tests

After installing in a clean environment, unittests were showing
differences of a few hundredths for some of the values.
To test this, I created a new test_solar.py module to check that
cza values from pvlib.spa match those of the new numpy-native
computations. This proved that values are within 12 or so decimal
places of precision of eachother.

Beyond that, I do not see anything that is drastically different
between the codes interms of solar adjustments and fraction of
direct radiation, so these very small differences in cza must
be propagating. I assume that the computations of cza are slighly
more accurate in our version as numpy.sum is used which may be
more accurate than the accumulator scalar that is used in pvlib
and in the old implementation.

Have updated some of the date handling to be more robust to errors,
though need to check this in other pipelines.

Have also forced copy of expanded coordinates when working with
Dataset objects to ensure that the values are contiguous so that
the call to .ravel() does not do weird things. Perhaps shoulds
add in checks/asserstions that must be contiguous.

* Dequantify DataArray after create

Ensure that the DataArray objects are dequantified after creation
to keep things simplier for downstream codes.

* Updates to handle keywords, datetime fix

Realized that if user inputs data values as keywords, the code for
checking if arguments are DataArrays was completely skipped.
So, added in an ARG_ORDER variable that is used to pop off keywords
in the proper order and append them to the args list.

Also added a bug fix for the min_speed value being converted to
a numpy array without units. In the except when trying to
reshape returned WGBT values, we check if object has 'units' attr,
and if does, then we re-add them after running np.asarray().

We now take the ".values" part of DatetimeIndex objects when
adjusting times.

* Updates to dimiceli test for x86

On x86 the test of factor_b for dimiceli was failing with differences
at the 7th decimal place. As these values are very large (e9), having
that much precision in the decimal is not required. We have reducted
the almost equal check to 5 decimals.

---------

Co-authored-by: Kyle R. Wodzicki <kwodzicki@cicsnc.org>