Moves solar.py to Cython
The numpy array broadcasting introduced in the solar.py module
could cause very large memory usage depending on the size of data
input. To fix this, and get around the numba issues, we have moved
the module to Cython, with the problematic portions of the code
converted to Cython and remaining functions left as python.

We have also combined some function calls to reduce the number of
variables floating around; e.g., there were places where a variable
was defined and only used in the call for another function.